Finished netlify functions
