+++ netlify.toml
@@ -1,2 +1,3 @@
 [build]
+  command = "npm run webpack:build"
   functions = "functions"
