+++ package.json
@@ -37,6 +37,7 @@
     "webpack-merge": "^5.1.1"
   },
   "dependencies": {
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "node-fetch": "^2.6.1"
   }
 }
