+++ src/index.js
@@ -1,12 +1,11 @@
 import './styles/index.scss';
 import StartModal from './scripts/startModal';
-import Game from './scripts/game';
 
 document.addEventListener("DOMContentLoaded", () => {
     const startModal = new StartModal();
     startModal.selectSpeed();
     startModal.startGame();
-    
+    debugger;
     setTimeout(() => {
         document.querySelector(".splash-bg").setAttribute("class", "hidden-splash-bg");
     }, 1000);
