+++ src/scripts/timer.js
